Settings attribute that uses scala.Long in a collection
